Begin restructuring, create base storer and loader classes, extract functionality
