+++ src/sparkx/Oscar.py
@@ -11,9 +11,10 @@ from sparkx.Particle import Particle
 from sparkx.Filter import *
 import numpy as np
 import os
+from sparkx.Loader.OscarLoader import OscarLoader
+from sparkx.Storer.BaseStorer import BaseStorer
 
-
-class Oscar:
+class Oscar(BaseStorer):
     """
     Defines an Oscar object.
 
@@ -89,44 +90,10 @@ class Oscar:
 
     Methods
     -------
-    particle_list:
-        Returns current Oscar data as nested list
-    particle_objects_list:
-        Returns current Oscar data as nested list of Particle objects
-    num_events:
-        Get number of events
-    num_output_per_event:
-        Get number of particles in each event
     oscar_format:
         Get Oscar format of the input file
-    particle_species:
-        Keep only particles with given PDG ids
-    remove_particle_species:
-        Remove particles with given PDG ids
-    participants:
-        Keep participants only
-    spectators:
-        Keep spectators only
-    lower_event_energy_cut:
-        Filters out events with total energy lower than a threshold.
-    charged_particles:
-        Keep charged particles only
-    uncharged_particles:
-        Keep uncharged particles only
-    strange_particles:
-        Keep strange particles only
     spacetime_cut:
         Apply spacetime cut to all particles
-    pT_cut:
-        Apply pT cut to all particles
-    rapidity_cut:
-        Apply rapidity cut to all particles
-    pseudorapidity_cut:
-        Apply pseudorapidity cut to all particles
-    spacetime_rapidity_cut:
-        Apply space-time rapidity cut to all particles
-    multiplicity_cut:
-        Apply multiplicity cut to all particles
     print_particle_lists_to_file:
         Print current particle data to file with same format
 
@@ -204,175 +171,13 @@ class Oscar:
     """
 
     def __init__(self, OSCAR_FILE, **kwargs):
-        """
-        Parameters
-        ----------
-        OSCAR_FILE : TYPE
-            DESCRIPTION.
-        **kwargs : TYPE
-            DESCRIPTION.
+        super().__init__(OSCAR_FILE,**kwargs)
+        self.oscar_format=self.loader_.oscar_format()
+        self.event_end_lines_ = self.loader_.event_end_lines()  
 
-        Raises
-        ------
-        TypeError
-            DESCRIPTION.
-
-        Returns
-        -------
-        None
-        """
-
-        if ".oscar" not in OSCAR_FILE:
-            raise FileNotFoundError(
-                "Input file is not in the OSCAR format. Input "
-                "file must have the ending .oscar"
-            )
-
-        self.PATH_OSCAR_ = OSCAR_FILE
-        self.oscar_format_ = None
-        self.num_output_per_event_ = None
-        self.num_events_ = None
-        self.particle_list_ = None
-        self.optional_arguments_ = kwargs
-        self.event_end_lines_ = []
-
-        for keys in self.optional_arguments_.keys():
-            if keys not in ["events", "filters"]:
-                raise ValueError("Unknown keyword argument used in constructor")
-
-        if "events" in self.optional_arguments_.keys() and isinstance(
-            self.optional_arguments_["events"], tuple
-        ):
-            self.__check_that_tuple_contains_integers_only(
-                self.optional_arguments_["events"]
-            )
-            if (
-                self.optional_arguments_["events"][0]
-                > self.optional_arguments_["events"][1]
-            ):
-                raise ValueError(
-                    "First value of event number tuple must be smaller than second value"
-                )
-            elif (
-                self.optional_arguments_["events"][0] < 0
-                or self.optional_arguments_["events"][1] < 0
-            ):
-                raise ValueError("Event numbers must be positive")
-        elif "events" in self.optional_arguments_.keys() and isinstance(
-            self.optional_arguments_["events"], int
-        ):
-            if self.optional_arguments_["events"] < 0:
-                raise ValueError("Event number must be positive")
-
-        self.set_oscar_format()
-        self.set_num_events()
-        self.set_num_output_per_event_and_event_footers()
-        self.set_particle_list(kwargs)
-
-    # PRIVATE CLASS METHODS
-    def __check_that_tuple_contains_integers_only(self, events_tuple):
-        """
-        Check if all elements inside the event tuple are integers.
-
-        Parameters
-        ----------
-        events_tuple : tuple
-            Tuple containing event boundary events for read in.
-
-        Raises
-        ------
-        TypeError
-            If one or more elements inside the event tuple are not integers.
-        """
-        if not all(isinstance(event, int) for event in events_tuple):
-            raise TypeError(
-                "All elements inside the event tuple must be integers."
-            )
-
-    def __get_num_skip_lines(self):
-        """
-        Get number of initial lines in Oscar file that are header or comment
-        lines and need to be skipped in order to read the particle output.
-
-        Returns
-        -------
-        skip_lines : int
-            Number of initial lines before data.
-
-        """
-        if (
-            not self.optional_arguments_
-            or "events" not in self.optional_arguments_.keys()
-        ):
-            skip_lines = 3
-        elif isinstance(self.optional_arguments_["events"], int):
-            if self.optional_arguments_["events"] == 0:
-                skip_lines = 3
-            else:
-                cumulate_lines = 0
-                for i in range(0, self.optional_arguments_["events"]):
-                    cumulate_lines += self.num_output_per_event_[i, 1] + 2
-                skip_lines = 3 + cumulate_lines
-        elif isinstance(self.optional_arguments_["events"], tuple):
-            line_start = self.optional_arguments_["events"][0]
-            if line_start == 0:
-                skip_lines = 3
-            else:
-                cumulate_lines = 0
-                for i in range(0, line_start):
-                    cumulate_lines += self.num_output_per_event_[i, 1] + 2
-                skip_lines = 3 + cumulate_lines
-        else:
-            raise TypeError(
-                'Value given as flag "events" is not of type '
-                + "int or a tuple of two int values"
-            )
-
-        return skip_lines
-
-    def __skip_lines(self, fname):
-        """
-        Once a file is opened with :code:`open()`, this method skips the
-        initial header and comment lines such that the first line called with
-        :code:`fname.readline()` is the first particle in the first event.
-
-        Parameters
-        ----------
-        fname : variable name
-            Name of the variable for the file opened with the :code:`open()`
-            command.
-
-        """
-        num_skip = self.__get_num_skip_lines()
-        for i in range(0, num_skip):
-            fname.readline()
-
-    def __get_num_read_lines(self):
-        if (
-            not self.optional_arguments_
-            or "events" not in self.optional_arguments_.keys()
-        ):
-            cumulated_lines = np.sum(self.num_output_per_event_, axis=0)[1]
-            # add number of comments
-            cumulated_lines += int(2 * len(self.num_output_per_event_))
-
-        elif isinstance(self.optional_arguments_["events"], int):
-            read_event = self.optional_arguments_["events"]
-            cumulated_lines = int(self.num_output_per_event_[read_event, 1] + 2)
-
-        elif isinstance(self.optional_arguments_["events"], tuple):
-            cumulated_lines = 0
-            event_start = self.optional_arguments_["events"][0]
-            event_end = self.optional_arguments_["events"][1]
-            for i in range(event_start, event_end + 1):
-                cumulated_lines += int(self.num_output_per_event_[i, 1] + 2)
-        else:
-            raise TypeError(
-                "Value given as flag events is not of type int or a tuple"
-            )
-
-        return cumulated_lines
-
+    def create_loader(self, OSCAR_FILE):
+        self.loader_= OscarLoader(OSCAR_FILE)
+    
     def __particle_as_list(self, particle):
         particle_list = []
         particle_list.append(float(particle.t))
@@ -422,246 +227,10 @@ class Oscar:
 
         return particle_list
 
-    def __update_num_output_per_event_after_filter(self):
-        for event in range(0, len(self.particle_list_)):
-            self.num_output_per_event_[event][1] = len(
-                self.particle_list_[event]
-            )
-
-    def __apply_kwargs_filters(self, event, filters_dict):
-        if not isinstance(filters_dict, dict) or len(filters_dict.keys()) == 0:
-            return event
-        for i in filters_dict.keys():
-            if i == "charged_particles":
-                if filters_dict["charged_particles"]:
-                    event = charged_particles(event)
-            elif i == "uncharged_particles":
-                if filters_dict["uncharged_particles"]:
-                    event = uncharged_particles(event)
-            elif i == "strange_particles":
-                if filters_dict["strange_particles"]:
-                    event = strange_particles(event)
-            elif i == "particle_species":
-                event = particle_species(
-                    event, filters_dict["particle_species"]
-                )
-            elif i == "remove_particle_species":
-                event = remove_particle_species(
-                    event, filters_dict["remove_particle_species"]
-                )
-            elif i == "participants":
-                if filters_dict["participants"]:
-                    event = participants(event)
-            elif i == "spectators":
-                if filters_dict["spectators"]:
-                    event = spectators(event)
-            elif i == "lower_event_energy_cut":
-                event = lower_event_energy_cut(
-                    event, filters_dict["lower_event_energy_cut"]
-                )
-            elif i == "spacetime_cut":
-                event = spacetime_cut(
-                    event,
-                    filters_dict["spacetime_cut"][0],
-                    filters_dict["spacetime_cut"][1],
-                )
-            elif i == "pT_cut":
-                event = pT_cut(event, filters_dict["pT_cut"])
-            elif i == "mT_cut":
-                event = mT_cut(event, filters_dict["mT_cut"])
-            elif i == "rapidity_cut":
-                event = rapidity_cut(event, filters_dict["rapidity_cut"])
-            elif i == "pseudorapidity_cut":
-                event = pseudorapidity_cut(
-                    event, filters_dict["pseudorapidity_cut"]
-                )
-            elif i == "spacetime_rapidity_cut":
-                event = spacetime_rapidity_cut(
-                    event, filters_dict["spacetime_rapidity_cut"]
-                )
-            elif i == "multiplicity_cut":
-                event = multiplicity_cut(
-                    event, filters_dict["multiplicity_cut"]
-                )
-            else:
-                raise ValueError("The cut is unknown!")
-
-        return event
-
-    # PUBLIC CLASS METHODS
-
-    def set_particle_list(self, kwargs):
-        particle_list = []
-        data = []
-        num_read_lines = self.__get_num_read_lines()
-        with open(self.PATH_OSCAR_, "r") as oscar_file:
-            self.__skip_lines(oscar_file)
-            for i in range(0, num_read_lines):
-                line = oscar_file.readline()
-                if not line:
-                    raise IndexError(
-                        "Index out of range of OSCAR file. This most likely happened because "
-                        + "the particle number specified by the comments in the OSCAR "
-                        + "file differs from the actual number of particles in the event."
-                    )
-                elif i == 0 and "#" not in line and "out" not in line:
-                    raise ValueError(
-                        "First line of the event is not a comment "
-                        + 'line or does not contain "out"'
-                    )
-                elif "event" in line and ("out" in line or "in " in line):
-                    continue
-                elif "#" in line and "end" in line:
-                    if "filters" in self.optional_arguments_.keys():
-                        data = self.__apply_kwargs_filters(
-                            [data], kwargs["filters"]
-                        )[0]
-                        self.num_output_per_event_[len(particle_list)] = (
-                            len(particle_list),
-                            len(data),
-                        )
-                    particle_list.append(data)
-                    data = []
-                elif "#" in line:
-                    raise ValueError("Comment line unexpectedly found: " + line)
-                else:
-                    line = line.replace("\n", "").split(" ")
-                    particle = Particle(self.oscar_format_, line)
-                    data.append(particle)
-
-        # Correct num_output_per_event and num_events
-        if not kwargs or "events" not in self.optional_arguments_.keys():
-            if len(particle_list) != self.num_events_:
-                raise IndexError(
-                    "Number of events in OSCAR file does not match the "
-                    + "number of events specified by the comments in the "
-                    + "OSCAR file!"
-                )
-        elif isinstance(kwargs["events"], int):
-            update = self.num_output_per_event_[kwargs["events"]]
-            self.num_output_per_event_ = [update]
-            self.num_events_ = int(1)
-        elif isinstance(kwargs["events"], tuple):
-            event_start = kwargs["events"][0]
-            event_end = kwargs["events"][1]
-            update = self.num_output_per_event_[event_start : event_end + 1]
-            self.num_output_per_event_ = update
-            self.num_events_ = int(event_end - event_start + 1)
-
-        if not kwargs or "events" not in self.optional_arguments_.keys():
-            self.particle_list_ = particle_list
-        elif isinstance(kwargs["events"], int):
-            self.particle_list_ = particle_list
-        else:
-            self.particle_list_ = particle_list
-
-    def set_oscar_format(self):
-        with open(self.PATH_OSCAR_, "r") as file:
-            first_line = file.readline()
-            first_line = first_line.replace("\n", "").split(" ")
-
-        if len(first_line) == 15 or first_line[0] == "#!OSCAR2013":
-            self.oscar_format_ = "Oscar2013"
-        elif (
-            first_line[0] == "#!OSCAR2013Extended"
-            and first_line[1] == "SMASH_IC"
-        ):
-            self.oscar_format_ = "Oscar2013Extended_IC"
-        elif (
-            first_line[0] == "#!OSCAR2013Extended"
-            and first_line[1] == "Photons"
-        ):
-            self.oscar_format_ = "Oscar2013Extended_Photons"
-        elif len(first_line) == 23 or first_line[0] == "#!OSCAR2013Extended":
-            self.oscar_format_ = "Oscar2013Extended"
-        else:
-            raise TypeError(
-                "Input file must follow the Oscar2013, "
-                + "Oscar2013Extended, Oscar2013Extended_IC or Oscar2013Extended_Photons format. "
-            )
-
-    def set_num_output_per_event_and_event_footers(self):
-        with open(self.PATH_OSCAR_, "r") as oscar_file:
-            event_output = []
-            if (
-                self.oscar_format_ != "Oscar2013Extended_IC"
-                and self.oscar_format_ != "Oscar2013Extended_Photons"
-            ):
-                while True:
-                    line = oscar_file.readline()
-                    if not line:
-                        break
-                    elif "#" in line and "end " in line:
-                        self.event_end_lines_.append(line)
-                    elif "#" in line and "out" in line:
-                        line_str = line.replace("\n", "").split(" ")
-                        event = line_str[2]
-                        num_output = line_str[4]
-                        event_output.append([event, num_output])
-                    else:
-                        continue
-            elif self.oscar_format_ == "Oscar2013Extended_IC":
-                line_counter = 0
-                event = 0
-                while True:
-                    line_counter += 1
-                    line = oscar_file.readline()
-                    if not line:
-                        break
-                    elif "#" in line and "end" in line:
-                        self.event_end_lines_.append(line)
-                        event_output.append([event, line_counter - 2])
-                    elif "#" in line and "in" in line:
-                        line_str = line.replace("\n", "").split(" ")
-                        event = line_str[2]
-                        line_counter = 0
-                    else:
-                        continue
-
-            elif self.oscar_format_ == "Oscar2013Extended_Photons":
-                line_counter = 0
-                event = 0
-                line_memory = 0
-                while True:
-                    line_counter += 1
-                    line_memory += 1
-                    line = oscar_file.readline()
-                    if not line:
-                        break
-                    elif "#" in line and "end" in line:
-                        if line_memory == 1:
-                            continue
-                        self.event_end_lines_.append(line)
-                        line_str = line.replace("\n", "").split(" ")
-                        event = line_str[2]
-                        event_output.append([event, line_counter - 1])
-                    elif "#" in line and "out" in line:
-                        line_counter = 0
-                    else:
-                        continue
-        self.num_output_per_event_ = np.asarray(event_output, dtype=np.int32)
-
-    def set_num_events(self):
-        # Read the file in binary mode to search for last line. In this way one
-        # does not need to loop through the whole file
-        with open(self.PATH_OSCAR_, "rb") as file:
-            file.seek(-2, os.SEEK_END)
-            while file.read(1) != b"\n":
-                file.seek(-2, os.SEEK_CUR)
-            last_line = file.readline().decode().split(" ")
-        if last_line[0] == "#" and "event" in last_line:
-            self.num_events_ = int(last_line[2]) + 1
-        else:
-            raise TypeError(
-                "Input file does not end with a comment line "
-                + "including the events. File might be incomplete "
-                + "or corrupted."
-            )
-
     def particle_list(self):
         """
         Returns a nested python list containing all quantities from the
-        current Oscar data as numerical values with the following shape:
+        current data as numerical values with the following shape:
 
             | Single Event:    [[output_line][particle_quantity]]
             | Multiple Events: [event][output_line][particle_quantity]
@@ -669,7 +238,7 @@ class Oscar:
         Returns
         -------
         list
-            Nested list containing the current Oscar data
+            Nested list containing the current data
 
         """
         num_events = self.num_events_
@@ -694,23 +263,7 @@ class Oscar:
                 particle_array.append(event)
 
         return particle_array
-
-    def particle_objects_list(self):
-        """
-        Returns a nested python list containing all particles from
-        the Oscar2013/Oscar2013Extended output as particle objects
-        from Particle:
-
-           | Single Event:    [particle_object]
-           | Multiple Events: [event][particle_object]
-
-        Returns
-        -------
-        particle_list_ : list
-            List of particle objects from Particle
-        """
-        return self.particle_list_
-
+    
     def oscar_format(self):
         """
         Get the Oscar format of the input file.
@@ -723,208 +276,6 @@ class Oscar:
 
         """
         return self.oscar_format_
-
-    def num_output_per_event(self):
-        """
-        Returns a numpy array containing the event number (starting with 1)
-        and the corresponding number of particles created in this event as
-
-        num_output_per_event[event_n, number_of_particles_in_event_n]
-
-        num_output_per_event is updated with every manipulation e.g. after
-        applying cuts.
-
-        Returns
-        -------
-        num_output_per_event_ : numpy.ndarray
-            Array containing the event number and the corresponding number of
-            particles
-        """
-        return self.num_output_per_event_
-
-    def num_events(self):
-        """
-        Returns the number of events in particle_list
-
-        num_events is updated with every manipulation e.g. after
-        applying cuts.
-
-        Returns
-        -------
-        num_events_ : int
-            Number of events in particle_list
-        """
-        return self.num_events_
-
-    def charged_particles(self):
-        """
-        Keep only charged particles in particle_list
-
-        Returns
-        -------
-        self : Oscar object
-            Containing charged particles in every event only
-        """
-
-        self.particle_list_ = charged_particles(self.particle_list_)
-
-        # for i in range(0, self.num_events_):
-        #     self.particle_list_[i] = [elem for elem in self.particle_list_[i]
-        # if (elem.charge != 0 and elem.charge != np.nan)]
-
-        # new_length = len(self.particle_list_[i])
-        # self.num_output_per_event_[i, 1] = new_length
-        self.__update_num_output_per_event_after_filter()
-        return self
-
-    def uncharged_particles(self):
-        """
-        Keep only uncharged particles in particle_list
-
-        Returns
-        -------
-        self : Oscar object
-            Containing uncharged particles in every event only
-        """
-
-        # for i in range(0, self.num_events_):
-        #     self.particle_list_[i] = [elem for elem in self.particle_list_[i]
-        #                                 if (elem.charge == 0 and elem.charge != np.nan)]
-        #     new_length = len(self.particle_list_[i])
-        #     self.num_output_per_event_[i, 1] = new_length
-
-        self.particle_list_ = uncharged_particles(self.particle_list_)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def strange_particles(self):
-        """
-        Keep only strange particles in particle_list
-
-        Returns
-        -------
-        self : Oscar object
-            Containing strange particles in every event only
-        """
-
-        self.particle_list_ = strange_particles(self.particle_list_)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def particle_species(self, pdg_list):
-        """
-        Keep only particle species given by their PDG ID in every event
-
-        Parameters
-        ----------
-        pdg_list : int
-            To keep a single particle species only, pass a single PDG ID
-
-        pdg_list : tuple/list/array
-            To keep multiple particle species, pass a tuple or list or array
-            of PDG IDs
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only particle species specified by pdg_list for every event
-
-        """
-
-        self.particle_list_ = particle_species(self.particle_list_, pdg_list)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def remove_particle_species(self, pdg_list):
-        """
-        Remove particle species from particle_list by their PDG ID in every
-        event
-
-        Parameters
-        ----------
-        pdg_list : int
-            To remove a single particle species only, pass a single PDG ID
-
-        pdg_list : tuple/list/array
-            To remove multiple particle species, pass a tuple or list or array
-            of PDG IDs
-
-        Returns
-        -------
-        self : Oscar object
-            Containing all but the specified particle species in every event
-
-        """
-
-        self.particle_list_ = remove_particle_species(
-            self.particle_list_, pdg_list
-        )
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def participants(self):
-        """
-        Keep only participants in particle_list
-
-        Returns
-        -------
-        self : Oscar object
-            Containing participants in every event only
-        """
-
-        self.particle_list_ = participants(self.particle_list_)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def spectators(self):
-        """
-        Keep only spectators in particle_list
-
-        Returns
-        -------
-        self : Oscar object
-            Containing spectators in every event only
-        """
-
-        self.particle_list_ = spectators(self.particle_list_)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def lower_event_energy_cut(self, minimum_event_energy):
-        """
-        Filters out events with total energy lower than a threshold.
-
-        Parameters
-        ----------
-        minimum_event_energy : int or float
-            The minimum event energy threshold. Should be a positive integer or float.
-
-        Returns
-        -------
-        self: Oscar object
-            The updated instance of the class contains only events above the
-            energy threshold.
-
-        Raises
-        ------
-        TypeError
-            If the minimum_event_energy parameter is not an integer or float.
-        ValueError
-            If the minimum_event_energy parameter is less than or equal to 0.
-        """
-
-        self.particle_list_ = lower_event_energy_cut(
-            self.particle_list_, minimum_event_energy
-        )
-        self.__update_num_output_per_event_after_filter()
-
-        return self
 
     def spacetime_cut(self, dim, cut_value_tuple):
         """
@@ -951,171 +302,6 @@ class Oscar:
 
         self.particle_list_ = spacetime_cut(
             self.particle_list_, dim, cut_value_tuple
-        )
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def pT_cut(self, cut_value_tuple):
-        """
-        Apply transverse momentum cut to all events by passing an acceptance
-        range by ::code`cut_value_tuple`. All particles outside this range will
-        be removed.
-
-        Parameters
-        ----------
-        cut_value_tuple : tuple
-            Tuple with the upper and lower limits of the pT acceptance
-            range :code:`(cut_min, cut_max)`. If one of the limits is not
-            required, set it to :code:`None`, i.e. :code:`(None, cut_max)`
-            or :code:`(cut_min, None)`.
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only particles complying with the transverse momentum
-            cut for all events
-        """
-
-        self.particle_list_ = pT_cut(self.particle_list_, cut_value_tuple)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def mT_cut(self, cut_value_tuple):
-        """
-        Apply transverse mass cut to all events by passing an acceptance
-        range by ::code`cut_value_tuple`. All particles outside this range will
-        be removed.
-
-        Parameters
-        ----------
-        cut_value_tuple : tuple
-            Tuple with the upper and lower limits of the mT acceptance
-            range :code:`(cut_min, cut_max)`. If one of the limits is not
-            required, set it to :code:`None`, i.e. :code:`(None, cut_max)`
-            or :code:`(cut_min, None)`.
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only particles complying with the transverse mass
-            cut for all events
-        """
-
-        self.particle_list_ = mT_cut(self.particle_list_, cut_value_tuple)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def rapidity_cut(self, cut_value):
-        """
-        Apply rapidity cut to all events and remove all particles with rapidity
-        not complying with cut_value
-
-        Parameters
-        ----------
-        cut_value : float
-            If a single value is passed, the cut is applied symmetrically
-            around 0.
-            For example, if cut_value = 1, only particles with rapidity in
-            [-1.0, 1.0] are kept.
-
-        cut_value : tuple
-            To specify an asymmetric acceptance range for the rapidity
-            of particles, pass a tuple (cut_min, cut_max)
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only particles complying with the rapidity cut
-            for all events
-        """
-
-        self.particle_list_ = rapidity_cut(self.particle_list_, cut_value)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def pseudorapidity_cut(self, cut_value):
-        """
-        Apply pseudo-rapidity cut to all events and remove all particles with
-        pseudo-rapidity not complying with cut_value
-
-        Parameters
-        ----------
-        cut_value : float
-            If a single value is passed, the cut is applied symmetrically
-            around 0.
-            For example, if cut_value = 1, only particles with pseudo-rapidity
-            in [-1.0, 1.0] are kept.
-
-        cut_value : tuple
-            To specify an asymmetric acceptance range for the pseudo-rapidity
-            of particles, pass a tuple (cut_min, cut_max)
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only particles complying with the pseudo-rapidity cut
-            for all events
-        """
-
-        self.particle_list_ = pseudorapidity_cut(self.particle_list_, cut_value)
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def spacetime_rapidity_cut(self, cut_value):
-        """
-        Apply space-time rapidity cut to all events and remove all particles
-        with space-time rapidity not complying with cut_value
-
-        Parameters
-        ----------
-        cut_value : float
-            If a single value is passed, the cut is applied symmetrically
-            around 0.
-            For example, if cut_value = 1, only particles with spatial rapidity
-            in [-1.0, 1.0] are kept.
-
-        cut_value : tuple
-            To specify an asymmetric acceptance range for the spatial rapidity
-            of particles, pass a tuple (cut_min, cut_max)
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only particles complying with the space-time rapidity cut
-            for all events
-        """
-
-        self.particle_list_ = spacetime_rapidity_cut(
-            self.particle_list_, cut_value
-        )
-        self.__update_num_output_per_event_after_filter()
-
-        return self
-
-    def multiplicity_cut(self, min_multiplicity):
-        """
-        Apply multiplicity cut. Remove all events with a multiplicity lower
-        than min_multiplicity
-
-        Parameters
-        ----------
-        min_multiplicity : int
-            Lower bound for multiplicity. If the multiplicity of an event is
-            lower than min_multiplicity, this event is discarded.
-
-        Returns
-        -------
-        self : Oscar object
-            Containing only events with a multiplicity >= min_multiplicity
-        """
-
-        self.particle_list_ = multiplicity_cut(
-            self.particle_list_, min_multiplicity
         )
         self.__update_num_output_per_event_after_filter()
